terraform/1_core/services/monitoring/9_custom_checks/inactive_accounts/function_source/requirements.txt
google-cloud-bigquery
google-cloud-monitoring
